Added update to local storage to Drag and Drop
there was problem with moving the function iside the same list,because the index of returned the first index so if you moved the element up the list it would delete the new task from the local storage insted of the old one.
the problem us fixed with adding the if(r.208) and the support function-findElementIndexInLocalStorage.